style: fix code formatting and whitespace issues
files changed:
- script/deploy/*.sol
- src/dragons/*.sol
- src/hats/*.sol
- src/interfaces/*.sol
- test/*.sol

standardizes code formatting across project files, removes trailing whitespace, adds missing newlines, normalizes import statements, and improves overall code readability through consistent indentation and bracket placement.

+++ src/hats/AbstractHatsManager.sol
@@ -14,15 +14,15 @@ abstract contract AbstractHatsManager is IHatsEligibility, IHatsToggle {
     IHats public immutable HATS;
 
     // Core hat structure
-    uint256 public immutable adminHat;     // Parent admin hat
-    uint256 public immutable branchHat;    // Branch admin hat
-    
+    uint256 public immutable adminHat; // Parent admin hat
+    uint256 public immutable branchHat; // Branch admin hat
+
     // Role management
-    mapping(bytes32 => uint256) public roleHats;      // Maps role identifiers to hat IDs
-    mapping(uint256 => bytes32) public hatRoles;      // Reverse mapping of hats to role identifiers
-    
+    mapping(bytes32 => uint256) public roleHats; // Maps role identifiers to hat IDs
+    mapping(uint256 => bytes32) public hatRoles; // Reverse mapping of hats to role identifiers
+
     bool public isActive = true;
-    
+
     event RoleHatCreated(bytes32 roleId, uint256 hatId);
     event RoleGranted(bytes32 roleId, address account, uint256 hatId);
     event RoleRevoked(bytes32 roleId, address account, uint256 hatId);
@@ -33,14 +33,10 @@ abstract contract AbstractHatsManager is IHatsEligibility, IHatsToggle {
      * @param _adminHat The admin hat ID that will have admin privileges
      * @param _branchHat The branch hat ID
      */
-    constructor(
-        address hats, 
-        uint256 _adminHat,
-        uint256 _branchHat
-    ) {
+    constructor(address hats, uint256 _adminHat, uint256 _branchHat) {
         require(hats != address(0), "Invalid Hats address");
         HATS = IHats(hats);
-        
+
         require(HATS.isWearerOfHat(msg.sender, _adminHat), "Deployer must wear admin hat");
         adminHat = _adminHat;
         branchHat = _branchHat;
@@ -63,26 +59,26 @@ abstract contract AbstractHatsManager is IHatsEligibility, IHatsToggle {
         require(HATS.isAdminOfHat(msg.sender, adminHat), "Not admin");
         require(roleHats[roleId] == 0, "Role already exists");
         require(initialHolders.length <= maxSupply, "Too many initial holders");
-        
+
         // Create role hat under branch
         hatId = HATS.createHat(
             branchHat,
             details,
             uint32(maxSupply),
-            address(this),  // this contract determines eligibility
-            address(this),  // this contract controls activation
-            true,          // can be modified by admin
-            ""            // no custom image
+            address(this), // this contract determines eligibility
+            address(this), // this contract controls activation
+            true, // can be modified by admin
+            "" // no custom image
         );
-        
+
         roleHats[roleId] = hatId;
         hatRoles[hatId] = roleId;
 
         // Mint hats to initial holders
-        for(uint i = 0; i < initialHolders.length; i++) {
+        for (uint256 i = 0; i < initialHolders.length; i++) {
             address holder = initialHolders[i];
             require(holder != address(0), "Invalid address");
-            
+
             // mintHat(
             //    uint256 hatId,    // Hat ID to mint
             //    address wearer    // Address to receive hat
@@ -90,7 +86,7 @@ abstract contract AbstractHatsManager is IHatsEligibility, IHatsToggle {
             HATS.mintHat(hatId, holder);
             emit RoleGranted(roleId, holder, hatId);
         }
-        
+
         emit RoleHatCreated(roleId, hatId);
     }
 
@@ -103,10 +99,10 @@ abstract contract AbstractHatsManager is IHatsEligibility, IHatsToggle {
     function grantRole(bytes32 roleId, address account) public virtual {
         require(HATS.isWearerOfHat(msg.sender, adminHat), "Not admin");
         require(account != address(0), "Invalid address");
-        
+
         uint256 hatId = roleHats[roleId];
         require(hatId != 0, "Role does not exist");
-        
+
         // Mint role hat
         HATS.mintHat(hatId, account);
         emit RoleGranted(roleId, account, hatId);
@@ -119,11 +115,11 @@ abstract contract AbstractHatsManager is IHatsEligibility, IHatsToggle {
      */
     function revokeRole(bytes32 roleId, address account) public virtual {
         require(HATS.isWearerOfHat(msg.sender, adminHat), "Not admin");
-        
+
         uint256 hatId = roleHats[roleId];
         require(hatId != 0, "Role does not exist");
         require(HATS.isWearerOfHat(account, hatId), "Address does not have role");
-        
+
         // Burn role hat
         HATS.setHatWearerStatus(hatId, account, false, false);
         emit RoleRevoked(roleId, account, hatId);
@@ -155,4 +151,4 @@ abstract contract AbstractHatsManager is IHatsEligibility, IHatsToggle {
         require(HATS.isWearerOfHat(msg.sender, adminHat), "Not admin");
         isActive = !isActive;
     }
-} 
+}
